requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
docstring_parser==0.16
google-ai-generativelanguage==0.6.9
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.63.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-generativeai==0.8.1
google-generativeai
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
jiter==0.7.0
numpy==2.1.0
openai>=1.50.0
packaging==24.1
pillow~=11.1.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv
python-telegram-bot==21.4
requests==2.32.3
rsa==4.9
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
vertexai==1.63.0
SQLAlchemy~=2.0.39
psycopg2-binary
